+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+  - Initial release
+  - `read` function to read .env files and return a map of key/value pairs